Split up integration tests (2/2)
As discussed in #656
